Rearranged lending instructions
