+++ src/index.ts
@@ -1,5 +1,5 @@
 export {Port} from './Port';
-export * from './instructions';
+export * from './instructions/lending';
 export * from './models';
 export * from './parsers';
 export * from './structs';
