--- src/instructions/index.ts
@@ -1,7 +1,0 @@
-export * from "./createLendingMarket";
-export * from "./createReserve";
-export * from "./refreshObligation";
-export * from "./refreshReserve";
-export * from "./liquidateObligation";
-export * from "./redeemReserveCollateral";
-export * from "./createStakingPool";
