+++ src/instructions/lending/depositObligationCollateral.ts
@@ -2,10 +2,10 @@ import { PublicKey, SYSVAR_CLOCK_PUBKEY, TransactionInstruction } from '@solana/
 import { TOKEN_PROGRAM_ID } from '@solana/spl-token';
 
 import * as BufferLayout from 'buffer-layout';
-import * as Layout from '../utils/layout';
+import * as Layout from 'src/utils/layout';
 
 import { LendingInstruction } from './instruction';
-import { AccessType, getAccess } from '../utils/Instructions';
+import { AccessType, getAccess } from 'src/utils/Instructions';
 
 // / Deposit collateral to an obligation. Requires a refreshed reserve.
 // /
