+++ src/instructions/lending/index.ts
@@ -1,0 +1,16 @@
+export * from "./borrowObligationLiquidity";
+export * from "./createLendingMarket";
+export * from "./createReserve";
+export * from "./createStakingPool";
+export * from "./depositObligationCollateral";
+export * from "./depositReserveLiquidity";
+export * from "./depositReserveLiquidityAndAddCollateral";
+export * from "./initObligation";
+export * from "./liquidateObligation";
+export * from "./redeemReserveCollateral";
+export * from "./refreshObligation";
+export * from "./refreshReserve";
+export * from "./repayObligationLiquidity";
+export * from "./updateReserve";
+export * from "./withdrawFee";
+export * from "./withdrawObligationCollateral";
