+++ src/instructions/lending/redeemReserveCollateral.ts
@@ -5,10 +5,10 @@ import {
 } from '@solana/web3.js';
 import BN from 'bn.js';
 import * as BufferLayout from 'buffer-layout';
-import * as Layout from '../utils/layout';
+import * as Layout from 'src/utils/layout';
 import {TOKEN_PROGRAM_ID} from '@solana/spl-token';
 import {LendingInstruction} from './instruction';
-import {PORT_LENDING} from '../constants';
+import {PORT_LENDING} from 'src/constants';
 
 // / Redeem collateral from a reserve in exchange for liquidity.
 // /
