+++ src/instructions/lending/refreshReserve.ts
@@ -4,7 +4,7 @@ import {
   TransactionInstruction,
 } from "@solana/web3.js";
 import * as BufferLayout from "buffer-layout";
-import { PORT_LENDING } from "../constants";
+import { PORT_LENDING } from "src/constants";
 import { LendingInstruction } from "./instruction";
 
 interface Data {
