+++ src/instructions/lending/repayObligationLiquidity.ts
@@ -1,10 +1,10 @@
 import {PublicKey, SYSVAR_CLOCK_PUBKEY, TransactionInstruction,} from '@solana/web3.js';
 import {TOKEN_PROGRAM_ID} from '@solana/spl-token';
 import * as BufferLayout from 'buffer-layout';
-import * as Layout from '../utils/layout'
-import { Lamport } from '../models/Lamport';
+import * as Layout from 'src/utils/layout'
+import { Lamport } from 'src/models/Lamport';
 import { LendingInstruction } from './instruction';
-import { AccessType, getAccess } from '../utils/Instructions';
+import { AccessType, getAccess } from 'src/utils/Instructions';
 
 /// Repay borrowed liquidity to a reserve. Requires a refreshed obligation and reserve.
 ///
