+++ src/instructions/lending/updateReserve.ts
@@ -6,35 +6,36 @@ import {
 import { TOKEN_PROGRAM_ID } from "@solana/spl-token";
 import * as BufferLayout from "buffer-layout";
 import { LendingInstruction } from "./instruction";
-import { AccessType, getAccess } from "../utils/Instructions";
+import { AccessType, getAccess } from "src/utils/Instructions";
+import { ReserveConfig, ReserveConfigLayout } from "..";
 
-/// Withdraw fee from a reserve.
+/// Update configuration for an existing market reserve.
+///
 /// Accounts expected by this instruction:
 ///
-///   0. `[]` Reserve account
+///   0. `[writable]` Reserve account
 ///   1. `[]` Lending market account.
 ///   2. `[]` Derived lending market authority.
 ///   3. `[signer]` Lending market owner.
-///   4. `[writable]` Reserve fee account
-///   5. `[writable]` Destination fee account
-///   6. `[]` Rent sysvar.
-///   7. `[]` Token program id.
-export const withdrawFeeInstruction = (
+///   4. `[]` Rent sysvar.
+///   5. `[]` Token program id.
+export const updateReserveInstruction = (
   transaction: Transaction,
+  config: ReserveConfig,
   reservePubkey: PublicKey, // 0
   lendingMarketPubkey: PublicKey, // 1
   lendingMarketAuthorityPubkey: PublicKey, // 2
-  lendingMarketOwnerPubkey: PublicKey, // 3
-  reserveFeePubkey: PublicKey, // 4
-  dstFeePubkey: PublicKey, // 5
+  lendingMarketOwnerPubkey: PublicKey // 3
 ): TransactionInstruction => {
   const dataLayout = BufferLayout.struct([
     BufferLayout.u8("instruction"),
+    ReserveConfigLayout,
   ]);
   const data = Buffer.alloc(dataLayout.span);
   dataLayout.encode(
     {
-      instruction: LendingInstruction.WithdrawFee,
+      instruction: LendingInstruction.UpdateReserve,
+      config,
     },
     data
   );
@@ -43,8 +44,6 @@ export const withdrawFeeInstruction = (
     getAccess(lendingMarketPubkey, AccessType.READ),
     getAccess(lendingMarketAuthorityPubkey, AccessType.READ),
     getAccess(lendingMarketOwnerPubkey, AccessType.SIGNER),
-    getAccess(reserveFeePubkey, AccessType.WRITE),
-    getAccess(dstFeePubkey, AccessType.WRITE),
     getAccess(SYSVAR_RENT_PUBKEY, AccessType.READ),
     getAccess(TOKEN_PROGRAM_ID, AccessType.READ),
   ];
