+++ src/instructions/lending/withdrawObligationCollateral.ts
@@ -1,10 +1,10 @@
 import { PublicKey, SYSVAR_CLOCK_PUBKEY, TransactionInstruction } from '@solana/web3.js';
 import { TOKEN_PROGRAM_ID } from '@solana/spl-token';
 import * as BufferLayout from 'buffer-layout';
-import * as Layout from '../utils/layout'
+import * as Layout from 'src/utils/layout'
 import { LendingInstruction } from './instruction';
-import { Lamport } from '../models/Lamport';
-import { AccessType, getAccess } from '../utils/Instructions';
+import { Lamport } from 'src/models/Lamport';
+import { AccessType, getAccess } from 'src/utils/Instructions';
 
 /// Withdraw collateral from an obligation. Requires a refreshed obligation and reserve.
 ///
