+++ tsconfig.json
@@ -4,6 +4,10 @@
     "target": "es5",
     "module": "commonjs",
     "lib": ["es2015", "dom"],
+    "baseUrl": ".",
+    "paths": {
+      "src": ["src/"]
+    },
     "allowJs": true,
     "noImplicitAny": false,
     "sourceMap": true,
